refactor(client): scope Youtube player wrapper to TheMusicPlayer component
